requirements.txt
docx
flask
pdfminer.six
requests
